water coffee graphs now done
