+++ client/src/App.js
@@ -1,6 +1,6 @@
 import './App.css';
 import Dashboard from './Containers/Dashboard';
-import Graphs from './Components/Graphs';
+import Graphs from './Components/graphPageComponents/Graphs';
 import Diary from './Components/Diary';
 import NavBar from './Components/NavBar';
 import { BrowserRouter as Router, Routes, Route } from "react-router-dom";
